+++ animation_nodes/data_structures/default_lists/default_list.pyx
@@ -1,13 +1,13 @@
 cdef class DefaultList:
-    cdef long getRealLength(self):
+    cdef Py_ssize_t getRealLength(self):
         return 0
 
     @classmethod
     def getMaxLength(cls, *args):
         cdef:
             DefaultList defaultList
-            long maxLength = 0
-            long length
+            Py_ssize_t maxLength = 0
+            Py_ssize_t length
 
         for defaultList in args:
             length = defaultList.getRealLength()
